Move tests for components under gordo_components test dir
