+++ tests/gordo_components/builder/test_builder.py
