+++ tests/gordo_components/cli/test_cli.py
