+++ tests/gordo_components/data_provider/test_azure_utils.py
