+++ tests/gordo_components/data_provider/test_data_provider_datalake.py
