+++ tests/gordo_components/data_provider/test_data_providers.py
