+++ tests/gordo_components/dataset/test_filter_rows.py
