+++ tests/gordo_components/model/test_feedforward_autoencoder.py
