+++ tests/gordo_components/model/test_lstm_autoencoder.py
