+++ tests/gordo_components/model/test_model.py
