+++ tests/gordo_components/model/test_model_factories_utils.py
