+++ tests/gordo_components/model/test_register.py
