+++ tests/gordo_components/model/test_transformers.py
