+++ tests/gordo_components/serializer/test_serializer_from_definition.py
