+++ tests/gordo_components/serializer/test_serializer_into_definition.py
