+++ tests/gordo_components/server/test_gordo_server.py
