+++ tests/gordo_components/util/test_disk_registry.py
