+++ tests/gordo_components/watchman/test_watchman.py
